+++ MiniRPG/Assets/@Resources/Animations/Player/PlayerTests.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c089f3b1e45d9094e9ef36ab341bcd6a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
