+++ MiniRPG/Assets/Scripts/Controller/MonsterController.cs
@@ -105,11 +105,10 @@ public class MonsterController : MonoBehaviour
 
     public void MonsterAttacked(int damage) // 몬스터가 공격받는 부분 => 플레이어 AttackSystem에서 호출하면됌.
     {
-        damage = 10; // 임시 데미지
         curHealth -= (int)damage;
-        if (!isDead && !isAttacked)
+        Debug.Log(curHealth);
+        if (!isDead)
         {
-            isAttacked = true;
             if (curHealth > 0)
             {
                 Debug.Log("몬스터가 공격 받았습니다.");
@@ -122,7 +121,6 @@ public class MonsterController : MonoBehaviour
                 Die();
                 Debug.Log("몬스터가 죽었습니다.");
             }
-            Invoke("ResetAttacked", 0.8f);
         }
     }
