+++ MiniRPG/Assets/Scripts/Models/Player/PlayerAttack2.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3419b606267108f4b9493cdc4ecc010c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
